feat: Basic Grid System Complete & Camera Zoom (#6)
* feat: added player character and movement

* Initial commit for grid system

Added a TileMapLayer for the walls.

* Working grid system

Set up a couple TileMapLayers. One for the floor (white) and one for the walls (black). Used simple 16x16 all-black/white square images for the TileSheets. Added collision/physics on the walls. Also added a temporary block to demo the collision physics.

* feat: added health variable and input map for non-movement actions

* feat: removed extra mapping that are in other branch

* Temp file to test pushing from work PC

* Delete deleteMe.txt

* Integrated engineer and a couple extra test blocks

* Removed test cubes

* Interaction keybinds

* Globally interactive doors

Need to fix tilemaplayers closed door sprites overlaying on top of open doors

* TileMapLayer sprite overlay fix

* Added mouse scroll camera zoom

---------

Co-authored-by: typhus733 <typhus733@gmail.com>