feat: Object Interaction (#7)
* Created interaction area component

Attach this new component to anything that you want the player to be able to interact with. You'll have to manually add a collision shape to the component after attaching it to an object. This is intentional so you can customize the size/shape of the interactable area of the object (example: a button that can be interacted with on a machine will have a different area than the entirety of the machine itself.)

* Doors converted to individually interactable objects

A door now has to be within the interaction range of the player and has to be the closest interactable object in order to be opened.

* Player interaction viewcone