Idle Turn state now working
This state causes the npc to change their 'target' to a point close to them, but at a different angle (depending on the maxTurnDegrees) to what they're currently at, simulating looking around. Also fixed the initial look direction so entities with the targeting component start off looking the way you put them in the 2D scene editor.