feat: Extra Initial States (#20)
* Extra States

* Split out functionality to different states

* fix zombie spawning bug

* Idle Turn state now working

This state causes the npc to change their 'target' to a point close to them, but at a different angle (depending on the maxTurnDegrees) to what they're currently at, simulating looking around. Also fixed the initial look direction so entities with the targeting component start off looking the way you put them in the 2D scene editor.

* fix state transition bug

* Turning logic and cleanup npc states

Moved turning logic to the targeter component. Cleaned up the npc_idle_turn state

* Refactor components and complete idlewalk

Refactored targeter component to fix bug with multiple possible targets. Refactored line of sight to add overloading to it's main function. Fleshed out the npc idle walk state to actually walk in the direction the entity is looking. Still working on a bug where allied soldiers get stuck on walls during the walk state.

* Revert LineOfSight Overload

This reverts the LineOfSight part of commit 7f1ad24ccff44acaf794adb23474ed4355c50e72.

* NPCs walking minimum distance

* Cleanup

* Added basic states to zombies

Also refactored the npc idle walk script so it doesn't listen to the navAgent's signal and use up processing power when we're not even in that state.